+++ apps/designer/app/designer/features/style-panel/sections/spacing/input-popover.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { ComponentProps, useState } from "react";
 import {
   keyframes,
   Popover,
@@ -7,13 +7,14 @@ import {
   PopoverPortal,
   styled,
 } from "@webstudio-is/design-system";
+import { StyleUpdate } from "@webstudio-is/project";
+import type { StyleProperty, StyleValue } from "@webstudio-is/css-data";
 import {
   CssValueInput,
   type IntermediateStyleValue,
 } from "../../shared/css-value-input";
-import type { StyleProperty, StyleValue } from "@webstudio-is/css-data";
-import type { StyleChangeHandler } from "./types";
 import type { StyleSource } from "../../shared/style-info";
+import { StyleUpdateOptions } from "../../shared/use-style-data";
 
 const slideUpAndFade = keyframes({
   "0%": { opacity: 0, transform: "scale(0.8)" },
@@ -30,7 +31,7 @@ const Input = ({
   styleSource: StyleSource;
   property: StyleProperty;
   value: StyleValue;
-  onChange: StyleChangeHandler;
+  onChange: (update: StyleUpdate, options: StyleUpdateOptions) => void;
   onClosePopover: () => void;
 }) => {
   const [intermediateValue, setIntermediateValue] = useState<
@@ -112,7 +113,7 @@ export const InputPopover = ({
   styleSource: StyleSource;
   property: StyleProperty;
   value: StyleValue;
-  onChange: StyleChangeHandler;
+  onChange: ComponentProps<typeof Input>["onChange"];
   isOpen: boolean;
   onClose: () => void;
 }) => {
